Start work on JavaScript to parse and decorate.
See example in "signature.html".  The idea of this code is that it
will be called once a SRFI has been loaded, and that it will parse the
signature text in each <scheme-procedure> element, replacing it with
markup that makes the signature clearer and better-looking.  With
JavaScript disabled, the signature would just be rendered as plain
text.